move usecase to parent package
