+++ domain/src/main/java/com/keelim/domain/GetFavoriteListUseCase.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.keelim.domain.nandadiagnosis
+package com.keelim.domain
 
 import com.keelim.data.model.entity.NandaEntity
 import com.keelim.data.source.NandaIORepository
@@ -22,7 +22,6 @@ import javax.inject.Inject
 class GetFavoriteListUseCase @Inject constructor(
     private val nandaIoRepository: NandaIORepository,
 ) {
-
     suspend operator fun invoke(): List<NandaEntity> {
         return nandaIoRepository.getFavoriteList()
     }
